// src/pages/cart/line-item/enum.ts
enum TagNames {
  DIV = 'div',
  BUTTON = 'button',
  A = 'a',
  IMG = 'img',
  INPUT = 'input',
}

enum Styles {
  CONTAINER = 'line-item-wrapper',
  LINE_ITEM = 'line-item',
  IMG_LINK = 'line-item__img-link',
  IMG = 'line-item__img',
  LINK = 'line-item__link',
  ACTIONS_GROUP = 'line-item__actions-group',
  BTN_REMOVE = 'line-item__btn',
  BTN_PLUS = 'line-item__btn-plus',
  BTN_MINUS = 'line-item__btn-minus',
  BTN_DISABLE_MINUS = 'line-item__btn-minus--disable',
  BTN_DISABLE_PLUS = 'line-item__btn-plus--disable',
  INPUT_WRAPPER = 'line-item__input-wrapper',
  INPUT = 'line-item__input',
  PRICE_WRAPPER = 'line-item__price-wrapper',
  PRICE = 'line-item__price',
  PRICE_DISCOUNT = 'line-item__dicsount-price',
  PRICE_GROUP = 'line-item__price-group',
  TOTAL_PRICE_WRAPPER = 'line-item__total-price-wrapper',
  TOTAL_PRICE = 'line-item__total-price',
  TOTAL_DISCOUNT = 'line-item__total-discount-price',
  PRICE_DISABLE = 'line-item__price--disabled',
}

enum Content {
  REMOVE_BTN = 'Remove',
}

export { TagNames, Styles, Content };